tests/requirements.txt
pytest
pytest-kind
requests
requests-html
python-slugify
#pytest-parallel
